Revert "Add bin_assigned signal"
This reverts commit c46ff70ecc09894912261d4acf6e24597fd1204b.

We want to have some extra information conveyed with the bin assigned signal.
For example, it'd be useful to know which user did so (and we also want to store
that information in the database to display it).

However, to get that information, the place changing the bin needs to emit the
signal instead, as only that place knows about things like the user.