Fix migration
Django models shouldn't be imported in migrations, as otherwise, we get the
newest version of the model but with an old database.

This migration was generated by Django, where it tried to include the default
get_inbox() callable, thus importing the models.

This seems to work fine with sqlite, but breaks initial migrations when setting
up crashbin with a MySQL database.

Since we only expect this migration to run when a fresh database gets
initialized, just remove the default entirely.